requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.1.0
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.6
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
entrypoints==0.4
fasteners==0.14.1
filelock==3.6.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
greenlet==2.0.1
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==2.1.2
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
mysql-connector-python==8.0.31
mysqlclient==2.1.1
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
nose==1.3.7
notebook==6.4.8
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipenv==2022.12.19
platformdirs==2.5.1
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==3.12.4
ptyprocess==0.7.0
pudb==2022.1.2
py==1.10.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.0.0
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.21.0
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
reportlab==3.6.8
requests==2.25.1
scipy==1.8.0
SecretStorage==3.3.1
Send2Trash==1.8.1b0
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
SQLAlchemy==2.0.0
sympy==1.9
systemd-python==234
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.1.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.4.0
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
urwid==2.1.2
urwid-readline==0.13
usb-creator==0.3.7
uuid==1.30
virtualenv==20.13.0+ds
virtualenv-clone==0.5.7
wadllib==1.3.6
waitress==2.1.2
wcwidth==0.2.5
webencodings==0.5.1
webssh==1.6.0
Werkzeug==2.2.2
widgetsnbextension==2.0.0
WTForms==3.0.1
xdg==5
xkit==0.0.0
zipp==1.0.0
